requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.12.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.0.0
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async==0.6.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
billiard==4.2.0
blinker==1.6.3
boto3==1.34.40
botocore==1.34.40
Brlapi==0.8.3
bs4==0.0.1
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
chardet==4.0.0
charset-normalizer==3.3.2
choicelib==0.1.5
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
command-not-found==0.3
constantly==23.10.4
contourpy==1.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
cupshelpers==1.0
cycler==0.12.1
daphne==4.1.0
dbus-python==1.2.18
defer==1.0.6
discord.py==2.3.2
dispatcher==1.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==2.1.0
Django==5.0.2
django-ckeditor-5==0.2.12
django-filter==23.5
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-minio-storage==0.5.7
django-mptt==0.16.0
django-resized==1.0.2
django-rest-swagger==2.2.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docker==6.1.3
drf-yasg==1.21.7
duplicity==0.8.21
ed25519==1.5
emoji==2.8.0
exceptiongroup==1.1.3
fasteners==0.14.1
flake8==7.0.0
Flask==3.0.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
fonttools==4.44.0
frozenlist==1.4.0
future==0.18.2
geographiclib==2.0
geopy==2.4.1
gif==23.3.0
greenlet==3.0.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.20.2
httpx==0.24.1
hyperlink==21.0.0
idna==3.6
importlib-metadata==4.6.4
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==23.5.0
kiwisolver==1.4.5
kombu==5.3.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.3
macaroonbakery==1.3.1
magic-filter==1.0.12
Mako==1.1.3
MarkupSafe==2.1.3
matplotlib==3.8.1
mccabe==0.7.0
minio==7.2.3
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.7
multidict==6.0.4
netifaces==0.11.0
numpy==1.26.0
oauthlib==3.2.0
olefile==0.46
openapi-codec==1.3.2
packaging==23.2
pandas==2.1.1
paramiko==2.9.3
pexpect==4.8.0
pg==0.1
phone==0.4.3
phonenumbers==8.13.22
pillow==10.2.0
pluggy==1.3.0
prompt-toolkit==3.0.43
protobuf==3.12.4
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==2.5.2
pydantic_core==2.14.5
pyflakes==3.2.0
pygame==2.5.0
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
PyOpenGL==3.1.7
pyOpenSSL==24.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyTelegramBotAPI==4.14.0
pytest==7.4.3
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-telegram-bot==20.5
pytz==2024.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==6.0.1
rabbitmq==0.2.0
redis==5.0.2
referencing==0.33.0
reportlab==3.6.8
requests==2.31.0
router==0.1
rpds-py==0.18.0
s3transfer==0.10.0
scrapping==0.1.dev0
SecretStorage==3.3.1
service-identity==24.1.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.22
sqlparse==0.4.4
ssh-import-id==5.11
swagger-spec-validator==3.0.3
systemd-python==234
Token==1.8.6
toml==0.10.2
tomli==2.0.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.0.7
usb-creator==0.3.7
user-agent==0.1.10
vbml==1.1.post1
vine==5.1.0
vkbottle==4.3.12
vkbottle-types==5.131.146.16
vkwave==0.2.17
wadllib==1.3.6
watchfiles==0.18.1
wcwidth==0.2.13
websocket-client==1.6.4
Werkzeug==3.0.0
xdg==5
xkit==0.0.0
yarl==1.9.3
youtube-dl==2021.12.17
zipp==1.0.0
zope.interface==6.2
